requirements.txt
rasa-sdk==1.1
rasa==1.1.4
grakn-client==1.5.3
